LayoutTests/editing/style/remove-styled-element-with-style-span-live-range.html

<!DOCTYPE html><!-- webkit-test-runner [ LiveRangeSelectionEnabled=true ] -->
<html>
<body>
<div id="test" contenteditable>hello <em class="Apple-style-span">world</em> WebKit</div>
<script src="../../resources/dump-as-markup.js"></script>
<script>

Markup.description('This test ensures WebKit removes implicitly styled elements even if they had class="Apple-style-span".'
+ ' There should be no span below.');

var test = document.getElementById('test');
test.focus();
document.execCommand('SelectAll', false, null);
document.execCommand('Italic', false, null);

Markup.dump(test);

</script>
</body>
</html>
